% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogfktheme.r
\name{rogfk}
\alias{rogfk}
\title{rogfk}
\usage{
rogfk
}
\value{
Beautifully formatted plot
}
\description{
A minimal ggplot theme that employs the golden ratio to great effect.
Credit to Christopher Adolph (with help from Kai Ping Leung) for Faceting
this theme (faculty.washington.edu/cadolph). Tweaked by Marius Swane Wishman.
}
\examples{
ggplot2::ggplot(data = penguins, ggplot2::aes(x = year, y = body_mass)) +
  ggplot2::geom_point() +
  rogfk

}
