% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogfktheme-package.R
\docType{package}
\name{rogfktheme-package}
\alias{rogfktheme}
\alias{rogfktheme-package}
\title{rogfktheme: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MariusSwane/rogfktheme}
  \item Report bugs at \url{https://github.com/MariusSwane/rogfktheme/issues}
}

}
\author{
\strong{Maintainer}: Marius Swane Wishman \email{marius.swane.wishman@rogfk.no} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Marius Swane Wishman \email{marius.swane.wishman@rogfk.no} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]
}

}
\keyword{internal}
