% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_rogfk_c.r
\name{scale_color_rogfk_c}
\alias{scale_color_rogfk_c}
\title{scale_color_rogfk_c}
\usage{
scale_color_rogfk_c(...)
}
\value{
En streng med hex-farge-coder fra mørk til lys/hvit blå.
}
\description{
Rogaland Fylkeskommune fargepalette for "continuous" data, basert på fargene
dis, dag, profil, aften og natt. Brukes som scale_color_gradientn() fra
ggplot2 og tar de samme parameterene for uten colors, som gis fra før.
}
\examples{
ggplot2::ggplot(
  data = penguins,
  ggplot2::aes(x = bill_len, y = body_mass, color = flipper_len)
) +
  ggplot2::geom_point() +
  scale_color_rogfk_c()

}
