% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_rogfk_d.r
\name{scale_color_rogfk_d}
\alias{scale_color_rogfk_d}
\title{scale_color_rogfk_d}
\usage{
scale_color_rogfk_d(...)
}
\value{
Describe here what the function returns.
}
\description{
Rogaland Fylkeskommune fargepalette for "discrete" data, med 11 farger. Etter
de 6 første begynner det å bli "repetetivt", så det er anbefalt å ikke bruke
fler enn det, med mindre det er strengt tatt nødvendig.
}
\examples{
ggplot2::ggplot(data = penguins) +
  ggplot2::geom_line(
    ggplot2::aes(x = bill_len, y = body_mass, color = species)
  ) +
  scale_color_rogfk_d()

}
