% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_rogfk_c.r
\name{scale_fill_rogfk_c}
\alias{scale_fill_rogfk_c}
\title{scale_fill_rogfk_c}
\usage{
scale_fill_rogfk_c(...)
}
\value{
En streng med hex-farge-coder fra mørk til lys/hvit blå.
}
\description{
Rogaland Fylkeskommune fargepalette for "continuous" data, basert på fargene
dis, dag, profil, aften og natt. Brukes som scale_color_gradientn() fra
ggplot2 og tar de samme parameterene for uten fill, som gis fra før.
}
\examples{
ggplot2::ggplot(data = penguins) +
  ggplot2::geom_col(
    ggplot2::aes(x = species, y = body_mass, fill = flipper_len)
  ) +
  scale_fill_rogfk_c()

}
