% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogfktheme-package.R
\docType{package}
\name{rogfktheme-package}
\alias{rogfktheme}
\alias{rogfktheme-package}
\title{rogfktheme: RogFK Theme}
\description{
Et "theme" med fargepalett hentet fra/inspirert av fylkeskommunenes grafiske anbefalinger og ggplot theme med minimale og objektivt optimale instillinger.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MariusSwane/rogfktheme}
  \item Report bugs at \url{https://github.com/MariusSwane/rogfktheme/issues}
}

}
\author{
\strong{Maintainer}: Marius Swane Wishman \email{marius.swane.wishman@rogfk.no} [copyright holder]

Authors:
\itemize{
  \item Marius Swane Wishman \email{marius.swane.wishman@rogfk.no} [copyright holder]
}

}
\keyword{internal}
